# swaggers/dashboard-notice-faq-inquiry/swagger.yml
openapi: 3.0.0
info:
  title: 대시보드/공지사항/자주묻는질문
  version: 1.0.0
servers:
  - url: http://220.90.208.226:8080
...
paths:
...
  /master/notice:
    post:
      tags:
        - 공지사항
      summary: 공지 생성
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: test notice title
                content:
                  type: string
                  example: test notice content
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: success
                data:
                  id: 51
                  draft: 0
                  title: test notice title
                  content: test notice content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-20T04:44:35.797+00:00'
                  dateEdit: '2021-12-20T04:44:35.797+00:00'
    get:
      tags:
        - 공지사항
      summary: 공지 조회
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '46'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                id: 46
                draft: 0
                title: test notice title
                content: test notice content
                views: 0
                imageUrl: null
                dateAdd: '2021-12-13T07:57:36.659+00:00'
                dateEdit: '2021-12-13T07:57:36.659+00:00'
    put:
      tags:
        - 공지사항
      summary: 공지 수정
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                draft: 1
                title: new test title
                content: new test content
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '45'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                id: 45
                draft: 1
                title: new test title
                content: new test content
                views: 0
                imageUrl: null
                dateAdd: '2021-12-13T07:55:40.122+00:00'
                dateEdit: '2021-12-13T08:10:49.579+00:00'
    delete:
      tags:
        - 공지사항
      summary: 공지 삭제
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '4'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: notice deleted
  /master/notice/draft:
    post:
      tags:
        - 공지사항
      summary: 공지 임시저장
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: draft notice title
                content:
                  type: string
                  example: draft notice content
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: success
                data:
                  id: 52
                  draft: 1
                  title: draft notice title
                  content: draft notice content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-20T04:45:14.335+00:00'
                  dateEdit: '2021-12-20T04:45:14.335+00:00'
  /master/notice/all:
    get:
      tags:
        - 공지사항
      summary: 공지 전체 조회
      parameters:
        - name: search
          in: query
          schema:
            type: string
          example: title
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                - id: 45
                  draft: 1
                  title: new test title
                  content: new test content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-13T07:55:40.122+00:00'
                  dateEdit: '2021-12-13T08:10:49.579+00:00'
                - id: 46
                  draft: 0
                  title: test notice title
                  content: test notice content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-13T07:57:36.659+00:00'
                  dateEdit: '2021-12-13T07:57:36.659+00:00'
                - id: 48
                  draft: 1
                  title: test notice title
                  content: test notice content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-13T08:07:49.884+00:00'
                  dateEdit: '2021-12-13T08:07:49.884+00:00'
  /notice/image:
    get:
      tags:
        - 공지사항
      summary: 공지사항 이미지
      parameters:
        - name: fileName
          in: query
          schema:
            type: string
          example: e12df221-e765-4a9e-80a6-c90e6bb08e63_테스트 그림.jpg
      responses:
        '200':
          description: Successful response
  /master/faq:
    post:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 생성
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                category:
                  type: string
                  example: 이용문의
                title:
                  type: string
                  example: test faq title
                content:
                  type: string
                  example: test faq content
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: success
                data:
                  id: 33
                  category: 이용문의
                  draft: 0
                  title: test faq title
                  content: test faq content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-20T04:46:32.255+00:00'
                  dateEdit: '2021-12-20T04:46:32.255+00:00'
    get:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 조회
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '34'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                id: 34
                category: draft category
                draft: 1
                title: draft faq title
                content: draft faq content
                views: 0
                imageUrl: null
                dateAdd: '2021-12-20T04:47:32.564+00:00'
                dateEdit: '2021-12-20T04:47:32.564+00:00'
    put:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 수정
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                draft: 1
                category: new_c
                title: new faq title
                content: new faq content
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '33'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                id: 33
                category: new_c
                draft: 1
                title: new faq title
                content: new faq content
                views: 0
                imageUrl: null
                dateAdd: '2021-12-20T04:46:32.255+00:00'
                dateEdit: '2021-12-20T05:17:39.621+00:00'
    delete:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 삭제
      parameters:
        - name: id
          in: query
          schema:
            type: integer
          example: '33'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: faq deleted
  /master/faq/draft:
    post:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 임시저장
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                category:
                  type: string
                  example: draft category
                title:
                  type: string
                  example: draft faq title
                content:
                  type: string
                  example: draft faq content
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: success
                data:
                  id: 34
                  category: draft category
                  draft: 1
                  title: draft faq title
                  content: draft faq content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-20T04:47:32.564+00:00'
                  dateEdit: '2021-12-20T04:47:32.564+00:00'
  /master/faq/all:
    get:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 전체 조회
      parameters:
        - name: categories
          in: query
          schema:
            type: string
          description: 분류
        - name: search
          in: query
          schema:
            type: string
          description: 검색
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                - id: 34
                  category: draft category
                  draft: 1
                  title: draft faq title
                  content: draft faq content
                  views: 0
                  imageUrl: null
                  dateAdd: '2021-12-20T04:47:32.564+00:00'
                  dateEdit: '2021-12-20T04:47:32.564+00:00'
  /faq/image:
    get:
      tags:
        - 자주묻는질문
      summary: 자주묻는질문 이미지
      parameters:
        - name: fileName
          in: query
          schema:
            type: string
          example: e12df221-e765-4a9e-80a6-c90e6bb08e63_테스트 그림.jpg
      responses:
        '200':
          description: Successful response
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
...